Bug, Fixed Grid

+++ Engine/Source/Engine/Implementation/XInput/DirectXInput.cpp
@@ -203,12 +203,12 @@ int DirectXInput::GetDeltaY() const
 
 Vec2 DirectXInput::GetMousePosition()
 {
-    return Vec2(mouseX, mouseY);
+    return Vec2(static_cast<float>(mouseX), static_cast<float>(mouseY));
 }
 
 Vec2 DirectXInput::GetMouseDelta()
 {
-    return Vec2(lastX - mouseX, lastY - mouseY);
+    return Vec2(static_cast<float>(lastX) - mouseX, static_cast<float>(lastY) - mouseY);
 }
 
 int DirectXInput::GetMouseScrollDelta()
